[🚚rename]: layout 폴더 위치 이동
components/layouts
=>
src/layout
